Move CuratedContentAgent and OnwardsPicker to common
It's not just `article` that will need to use them, it is also `applications` in `InteractiveController` that renders interactive articles.
